add mandatory api key for routes

+++ LinkedNotes/src/main/java/ch/hearc/mbu/web/LinkController.java
@@ -1,9 +1,12 @@
 package ch.hearc.mbu.web;
 
 import ch.hearc.mbu.repository.link.Link;
+import ch.hearc.mbu.repository.user.User;
 import ch.hearc.mbu.service.link.LinkService;
 import ch.hearc.mbu.service.note.NoteService;
+import ch.hearc.mbu.web.helper.AuthentificationHelper;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -11,23 +14,35 @@ import org.springframework.web.bind.annotation.*;
 import java.util.Optional;
 
 @RestController
-@RequestMapping(value = "/links")
+@RequestMapping(value = "/{api_key}/links")
 public class LinkController {
+    //TODO only let the user access his own links
+    String apiKey;
     @Autowired
     LinkService linkService;
     @Autowired
     NoteService noteService;
+    @Autowired
+    AuthentificationHelper authentificationHelper;
 
     @GetMapping(value = "/")
-    public ResponseEntity<Iterable<Link>> getLinks() {
-        //TODO get user from token
+    public ResponseEntity<Iterable<Link>> getLinks(@PathVariable String api_key) {
+        User user = authentificationHelper.getUserFromApiKey(api_key);
+        if(user == null)
+        {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
         Iterable<Link> links = linkService.getLinks();
         return ResponseEntity.ok().contentType(MediaType.APPLICATION_JSON).body(links);
     }
 
     @GetMapping(value = "/{id}")
-    public ResponseEntity<Link> getLink(@PathVariable long id) {
-        //TODO get user from token
+    public ResponseEntity<Link> getLink(@PathVariable long id, @PathVariable String api_key) {
+        User user = authentificationHelper.getUserFromApiKey(api_key);
+        if(user == null)
+        {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
         Optional<Link> link = linkService.getLink(id);
         if(link.isPresent())
         {
@@ -39,14 +54,17 @@ public class LinkController {
     //TODO
 //    @GetMapping(value = "/note/{id}")
 //    public ResponseEntity<Iterable<Link>> getLinksOfNoteId(@PathVariable long id) {
-//        //TODO get user from token
 //        Iterable<Link> links = linkService.getLinksOfNoteId(id);
 //        return ResponseEntity.ok().contentType(MediaType.APPLICATION_JSON).body(links);
 //    }
 
     @PostMapping(value = "/", consumes = "application/json")
-    public ResponseEntity<String> addLink(@RequestBody Link link) {
-        //TODO get user from token
+    public ResponseEntity<String> addLink(@RequestBody Link link, @PathVariable String api_key) {
+        User user = authentificationHelper.getUserFromApiKey(api_key);
+        if(user == null)
+        {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
         if (link.getType() == null || link.getName() == null || link.getNote1()== null || link.getNote2() == null) {
             return ResponseEntity.badRequest().body("Title, name or one of the note is null");
         }
@@ -56,8 +74,12 @@ public class LinkController {
     }
 
     @PutMapping(value = "/", consumes = "application/json")
-    public ResponseEntity<String> updateLink(@RequestBody Link link) {
-        //TODO get user from token
+    public ResponseEntity<String> updateLink(@RequestBody Link link, @PathVariable String api_key) {
+        User user = authentificationHelper.getUserFromApiKey(api_key);
+        if(user == null)
+        {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
         if (!linkService.idExists(link.getId()) || link.getType() == null || link.getName() == null || link.getNote1()== null || link.getNote2() == null) {
             return ResponseEntity.badRequest().body(null);
         }
@@ -72,8 +94,12 @@ public class LinkController {
     }
 
     @DeleteMapping(value = "/{id}")
-    public ResponseEntity<String> deleteLink(@PathVariable long id) {
-        //TODO get user from token
+    public ResponseEntity<String> deleteLink(@PathVariable long id, @PathVariable String api_key) {
+        User user = authentificationHelper.getUserFromApiKey(api_key);
+        if(user == null)
+        {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
         if(linkService.idExists(id)){
             linkService.deleteLink(id);
             return ResponseEntity.ok().contentType(MediaType.APPLICATION_JSON).body("{\"id\": " + id + "}");
