+++ LinkedNotes/src/main/java/ch/hearc/mbu/web/NoteController.java
@@ -1,8 +1,10 @@
 package ch.hearc.mbu.web;
 
 import ch.hearc.mbu.repository.note.Note;
+import ch.hearc.mbu.repository.user.User;
 import ch.hearc.mbu.service.note.NoteService;
 import ch.hearc.mbu.service.user.UserService;
+import ch.hearc.mbu.web.helper.AuthentificationHelper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
@@ -12,23 +14,33 @@ import org.springframework.web.bind.annotation.*;
 import java.util.Optional;
 
 @RestController
-@RequestMapping(value = "/notes")
+@RequestMapping(value = "/{api_key}/notes")
 public class NoteController {
     @Autowired
     NoteService noteService;
     @Autowired
     UserService userService;
-
+    @Autowired
+    AuthentificationHelper authentificationHelper;
+    //TODO only let the user access his own notes
     @GetMapping(value = "/")
-    public ResponseEntity<Iterable<Note>> getNotesOfUser() {
-        //TODO get user from token
+    public ResponseEntity<Iterable<Note>> getNotesOfUser(@PathVariable String api_key) {
+        User user = authentificationHelper.getUserFromApiKey(api_key);
+        if(user == null)
+        {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
         Iterable<Note> notes = noteService.getNotes();
         return ResponseEntity.ok().contentType(MediaType.APPLICATION_JSON).body(notes);
     }
 
     @GetMapping(value = "/{id}")
-    public ResponseEntity<Note> getNote(@PathVariable long id) {
-        //TODO get user from token
+    public ResponseEntity<Note> getNote(@PathVariable long id, @PathVariable String api_key) {
+        User user = authentificationHelper.getUserFromApiKey(api_key);
+        if(user == null)
+        {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
         Optional<Note> note = noteService.getNote(id);
         if(note.isPresent())
         {
@@ -38,8 +50,12 @@ public class NoteController {
     }
 
     @PostMapping(value = "/", consumes = "application/json")
-    public ResponseEntity<String> addNoteOfUser(@RequestBody Note note) {
-        //TODO get user from token
+    public ResponseEntity<String> addNoteOfUser(@RequestBody Note note, @PathVariable String api_key) {
+        User user = authentificationHelper.getUserFromApiKey(api_key);
+        if(user == null)
+        {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
         if (note.getTitle() == null || note.getContent() == null) {
             return ResponseEntity.badRequest().body("Title or content is null");
         }
@@ -49,8 +65,12 @@ public class NoteController {
     }
 
     @PutMapping(value = "/", consumes = "application/json")
-    public ResponseEntity<String> updateNote(@RequestBody Note note) {
-        //TODO get user from token
+    public ResponseEntity<String> updateNote(@RequestBody Note note, @PathVariable String api_key) {
+        User user = authentificationHelper.getUserFromApiKey(api_key);
+        if(user == null)
+        {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
         if (!noteService.idExists(note.getId()) || note.getTitle() == null || note.getContent() == null) {
             return ResponseEntity.badRequest().body(null);
         }
@@ -66,8 +86,12 @@ public class NoteController {
     }
 
     @DeleteMapping(value = "/{id}")
-    public ResponseEntity<String> deleteNoteOfUser(@PathVariable long id) {
-        //TODO get user from token
+    public ResponseEntity<String> deleteNoteOfUser(@PathVariable long id, @PathVariable String api_key) {
+        User user = authentificationHelper.getUserFromApiKey(api_key);
+        if(user == null)
+        {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
         if(noteService.idExists(id)){
             noteService.deleteNote(id);
             return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
